actualice todo el codigo y ahora ya esta todo aca (redux + router + api services)
