+++ src/app/screens/Home/styles.module.scss
@@ -1,41 +1,26 @@
 @import 'scss/variables/_mediaQueries.scss';
 
+.link {
+  position: absolute;
+  top: 0%;
+  left: 2%;
+  font-size: 20px;
+  color: black;
+}
+
+.link:hover {
+  font-size: 22px;
+  color: blue;
+  text-decoration: none;
+}
+
 .container {
   max-height: 100%;
   align-items: center;
-  @media #{$desktop}, #{$tablet} {
-    padding: 3% 0;
-  }
-
   @media #{$mobile} {
-    padding: 10% 5%;
     display: flex;
     flex-direction: column-reverse;
   }
-}
-
-.containerTrace {
-  @media #{$desktop}, #{$tablet} {
-    background-color: rgba(200, 200, 200, 1);
-    border-radius: 25px;
-    color: white;
-    width: 25%;
-    float: left;
-    display: flex;
-    justify-content: center;
-  }
-  @media #{$mobile} {
-    background-color: rgba(200, 200, 200, 1);
-    border-radius: 25px;
-    color: white;
-    width: 80%;
-    display: flex;
-    justify-content: center;
-  }
-}
-
-.trace {
-  background-color: white;
 }
 
 .containerCalculator {
