+++ src/app/screens/Trace/styles.module.scss
@@ -1,14 +1,23 @@
 @import 'scss/variables/_mediaQueries.scss';
 
+.link {
+  position: absolute;
+  top: 0%;
+  left: 2%;
+  font-size: 20px;
+  color: black;
+}
+
+.link:hover {
+  font-size: 22px;
+  color: blue;
+  text-decoration: none;
+}
+
 .container {
   max-height: 100%;
   align-items: center;
-  @media #{$desktop}, #{$tablet} {
-    padding: 3% 0;
-  }
-
   @media #{$mobile} {
-    padding: 10% 5%;
     display: flex;
     flex-direction: column-reverse;
   }
