+++ src/constants/routes.js
@@ -1,1 +1,2 @@
-
+export const HOME = '/';
+export const TRACE = '/trace';
