+++ src/scss/index.scss
@@ -16,7 +16,6 @@ body {
 
 #root {
   height: 100%;
-  padding: 3% 0;
   display: flex;
   align-items: center;
   flex-flow: column;
